random_chi's second argument can now be int or long.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11971 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -7081,7 +7081,7 @@ void NclAddUserFuncs(void)
 
     dimsizes[0] = 1;
     SetArgTemplate(args, nargs, "numeric", 1, dimsizes);  nargs++;
-    SetArgTemplate(args, nargs, "integer", 1, NclANY);  nargs++;
+    SetArgTemplate(args, nargs, "numeric", 1, NclANY);  nargs++;
 
     NclRegisterFunc(random_chi_W, args, "random_chi", nargs);
